Do not restrict profile argument
There are many correct choices for the profile argument. Restricting
the argument for the `--profile` argument in the parser is challenging
to maintain since it might consist of a product of profiles and
architectures.

This change removes any restriction on `--profile`. Note that this
means that a non-existing choice will fail in the container and might
be harder to debug.

Fixes: nicolasbock/ebuildtester#152
Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>